requirements.txt
pandas
openpyxl
requests
Pillow
python-dotenv
cloudinary
